[#1052] Fix the docker repository URL in case of Azure and AWS
Before these changes, we passed the docker URL parameter from
Jenkins's job into the setup robot tests script. This URL is used for
deploying pods inside k8s for testing purposes. It follows that in the
case of AWS and Azure, we must pass the docker URL from hiera because
the Docker URL in hiera doesn't contain `-local` suffix.